Alright, creating the client is almost done. Followed the example, found a different way of doing things with using server = gethostbyadd, not using bzero or bcopy and instead using memset and memcpy, and needed to declare a socklen_t to use with the memset.
Really hope this is working!

Next thing to do is provide cmd line arguments and see if things break.

If it works, we can create the server.

After creating the server, then we need to implement the stop and wait algorithm by replacing the connect and recieve portions of the client/server example.

You may want to confirm with the TA and work with the professor to confirm this a bit more...